+++ src/main/java/baseball/config/Constant.java
@@ -1,0 +1,12 @@
+package baseball.config;
+
+public class Constant {
+    public static final int NUM_SIZE = 3;
+    public static final int MIN_NUM_VALUE = 1;
+    public static final int MAX_NUM_VALUE = 9;
+    public static final int MAX_STRIKES = 3;
+    public static final int NONE = 0;
+    public static final int RESTART_GAME = 1;
+    public static final int END_GAME = 2;
+    public static final int INVALID_INDEX = -1;
+}
